refactor: Add metdata info to activity items UI
