+++ lib/net/queries/actions/useGetUserActivity.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 import client from "@/lib/net/client";
-import { UserActivity } from "./type";
 import { QueryHookOptions } from "@/lib/types/query";
+import { IUserActivity } from "@/lib/types/user";
 
 interface GetUserActivityPathParams {
   userId: string;
@@ -9,8 +9,8 @@ interface GetUserActivityPathParams {
 
 const buildUserActivityQueryKey = (params: GetUserActivityPathParams) => ["user-activity", params];
 
-const useGetUserActivity = ({ userId }: GetUserActivityPathParams, options?: QueryHookOptions<UserActivity[]>) => {
-  return useQuery<UserActivity[], Error>({
+const useGetUserActivity = ({ userId }: GetUserActivityPathParams, options?: QueryHookOptions<IUserActivity[]>) => {
+  return useQuery<IUserActivity[], Error>({
     queryKey: buildUserActivityQueryKey({ userId }),
     queryFn: () => getUserActivity(userId),
     enabled: !!userId,
@@ -19,7 +19,7 @@ const useGetUserActivity = ({ userId }: GetUserActivityPathParams, options?: Que
 };
 
 async function getUserActivity(userId: string) {
-  const response = await client.get<UserActivity[]>(`users/id/${userId}/activities`);
+  const response = await client.get<IUserActivity[]>(`users/id/${userId}/activities`);
   return response.data;
 }
