+++ lib/store/activityStore.ts
@@ -1,12 +1,12 @@
 import { create } from 'zustand';
 import { persist, createJSONStorage } from 'zustand/middleware';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import { UserActivity } from '@/lib/net/queries/actions/type';
+import { IUserActivity } from '../types/user';
 
 interface ActivityState {
-  activities: UserActivity[];
-  setActivities: (activities: UserActivity[]) => void;
-  addActivity: (activity: UserActivity) => void;
+  activities: IUserActivity[];
+  setActivities: (activities: IUserActivity[]) => void;
+  addActivity: (activity: IUserActivity) => void;
   clearActivities: () => void;
 }
 
@@ -14,8 +14,8 @@ export const useActivityStore = create<ActivityState>()(
   persist(
     (set) => ({
       activities: [],
-      setActivities: (activities: UserActivity[]) => set({ activities }),
-      addActivity: (activity: UserActivity) =>
+      setActivities: (activities: IUserActivity[]) => set({ activities }),
+      addActivity: (activity: IUserActivity) =>
         set((state) => ({ activities: [...state.activities, activity] })),
       clearActivities: () => set({ activities: [] }),
     }),
